requirements.txt
fastapi==0.116.0
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.95.1
azure-core==1.34.0
azure-search-documents==11.5.3
azure-cognitiveservices-speech
azure-storage-blob==12.17.0
pydantic-settings==2.10.1
python-multipart==0.0.20
langchain==0.3.26
langchain-community==0.3.27
langchain-openai==0.3.27
azure-identity==1.15.0
anthropic==0.61.0
aiomysql
weasyprint
pydub
PyPDF2

# Windows WeasyPrint issue
# pip install --upgrade --force-reinstall cffi fonttools 